# Do not generate extra load for ++/-- for maps/variables