I noticed that for ++/-- operators we generated extra
lookups and loads for maps and variables specifically.

Like for example for following program:

```
  tracepoint:syscalls:sys_enter_write
  {
    @write++;
  }
```

We should generate single map lookup followed by value
increase and map update, instead we generate 2 lookups:

```
   0: (b7) r6 = 0
   1: (7b) *(u64 *)(r10 -8) = r6
   2: (18) r1 = map[id:137]
   4: (bf) r2 = r10
   5: (07) r2 += -8
   6: (85) call __htab_map_lookup_elem#99808  <--- 1.lookup
   7: (15) if r0 == 0x0 goto pc+1
   8: (07) r0 += 56
   9: (7b) *(u64 *)(r10 -16) = r6
  10: (18) r1 = map[id:137]
  12: (bf) r2 = r10
  13: (07) r2 += -16
  14: (85) call __htab_map_lookup_elem#99808  <--- 2.lookup
  15: (15) if r0 == 0x0 goto pc+1
  16: (07) r0 += 56
  17: (b7) r1 = 1
  18: (15) if r0 == 0x0 goto pc+2
  19: (79) r1 = *(u64 *)(r0 +0)
  20: (07) r1 += 1
  21: (7b) *(u64 *)(r10 -8) = r1
  22: (18) r1 = map[id:137]
  24: (bf) r2 = r10
  25: (07) r2 += -16
  26: (bf) r3 = r10
  27: (07) r3 += -8
  28: (b7) r4 = 0
  29: (85) call htab_map_update_elem#104752
  30: (b7) r0 = 0
  31: (95) exit

```
Please notice the '1.lookup' result is discarded and '2.lookup'
takes over.

It's because for ++/-- we generate the lookup within the
code generation so the accept call is superfluous and generates
unneeded code.

With the fix the code is with single lookup:

```
   0: (b7) r1 = 0
   1: (7b) *(u64 *)(r10 -16) = r1
   2: (18) r1 = map[id:139]
   4: (bf) r2 = r10
   5: (07) r2 += -16
   6: (85) call __htab_map_lookup_elem#99808   <--- lookup
   7: (15) if r0 == 0x0 goto pc+1
   8: (07) r0 += 56
   9: (b7) r1 = 1
  10: (15) if r0 == 0x0 goto pc+2
  11: (79) r1 = *(u64 *)(r0 +0)
  12: (07) r1 += 1
  13: (7b) *(u64 *)(r10 -8) = r1
  14: (18) r1 = map[id:139]
  16: (bf) r2 = r10
  17: (07) r2 += -16
  18: (bf) r3 = r10
  19: (07) r3 += -8
  20: (b7) r4 = 0
  21: (85) call htab_map_update_elem#104752
  22: (b7) r0 = 0
  23: (95) exit
```
It's not big deal for Varibles case, because the double
load was optimized out:

```
  Before optimization
  -------------------
  define i64 @"tracepoint:syscalls:sys_enter_write"(i8*) section "s_tracepoint:syscalls:sys_enter_write_1" {
  entry:
    %printf_args = alloca %printf_t
    %"$local" = alloca i64
    %1 = bitcast i64* %"$local" to i8*
    call void @llvm.lifetime.start.p0i8(i64 -1, i8* %1)
    store i64 0, i64* %"$local"
    %get_ns = call i64 inttoptr (i64 5 to i64 ()*)()
    %2 = bitcast i64* %"$local" to i8*
    call void @llvm.lifetime.start.p0i8(i64 -1, i8* %2)
    store i64 %get_ns, i64* %"$local"
    %3 = load i64, i64* %"$local"    <--- 1.load
    %4 = load i64, i64* %"$local"    <--- 2.load
  ...
```

It's also now a little bit faster, checked with following command:

`  # perf stat -r 5 bpftrace -e 'tracepoint:syscalls:sys_enter_write { @write++; }' -c 'perf bench sched messaging -l 20000'`

Before:

  ```
Performance counter stats for 'bpftrace -e tracepoint:syscalls:sys_enter_write { @write++; } -c perf bench sched messaging -l 20000' (5 runs):

       2,392,396.71 msec task-clock                #   39.686 CPUs utilized            ( +-  0.68% )
        141,660,095      context-switches          #    0.059 M/sec                    ( +-  0.68% )
         53,586,928      cpu-migrations            #    0.022 M/sec                    ( +-  2.05% )
             46,323      page-faults               #    0.019 K/sec                    ( +-  0.47% )
  5,965,553,694,245      cycles                    #    2.494 GHz                      ( +-  0.68% )
  1,380,288,001,056      instructions              #    0.23  insn per cycle           ( +-  0.30% )
    295,619,937,390      branches                  #  123.566 M/sec                    ( +-  0.31% )
      4,498,686,547      branch-misses             #    1.52% of all branches          ( +-  0.13% )

             60.284 +- 0.409 seconds time elapsed  ( +-  0.68% )


```
After:
```
   Performance counter stats for 'bpftrace -e tracepoint:syscalls:sys_enter_write { @write++; } -c perf bench sched messaging -l 20000' (5 runs):

        2,193,764.69 msec task-clock                #   39.641 CPUs utilized            ( +-  0.98% )
         142,394,795      context-switches          #    0.065 M/sec                    ( +-  0.64% )
          54,758,486      cpu-migrations            #    0.025 M/sec                    ( +-  1.93% )
              46,414      page-faults               #    0.021 K/sec                    ( +-  0.13% )
   5,469,947,744,340      cycles                    #    2.493 GHz                      ( +-  0.98% )
   1,348,131,458,957      instructions              #    0.25  insn per cycle           ( +-  0.19% )
     288,773,429,973      branches                  #  131.634 M/sec                    ( +-  0.20% )
       4,487,594,062      branch-misses             #    1.55% of all branches          ( +-  0.32% )

              55.341 +- 0.519 seconds time elapsed  ( +-  0.94% )
```